phoenixcoder-oidc-server/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-jose[cryptography]==3.5.0
httpx==0.28.1
python-dotenv==1.1.1
psycopg2-binary==2.9.10
redis==6.2.0